# issue#27482 - added a check for if obj is instance of type in _isna-new