--- a/doc/source/whatsnew/v0.25.1.rst
+++ b/doc/source/whatsnew/v0.25.1.rst
@@ -93,7 +93,9 @@ Indexing
 Missing
 ^^^^^^^
 
--
+- Bug in :func:`pandas.isnull` or :func:`pandas.isna` when the input is a type e.g. `type(pandas.Series()) (:issue:27482)
+
+
 -
 -