@joy-rosie 2019-07-30
Can someone help me; when I run `git diff upstream/master -u -- "*.py" | flake8 --diff`
I get: `fatal: bad revision 'upstream/master'`