@jbrockmendel 2019-07-30
Needs a test, don’t worry too much about the git thing