@joy-rosie 2019-07-31
> this logic needs to be replicated in _isnew_old as well. tests are needed.
> 
> may want to move this to the last clause of the if/else. performance tests need to be run for this.

I added it above this elif https://github.com/joy-rosie/pandas/blob/issue-27482-isnull/pandas/core/dtypes/missing.py#L138 becasue it is that elif that is causing an error. I will add to _isna_old and also add tests, not sure how to add perfomance tests.