@joy-rosie 2019-08-01
I have added the tests and the statement to `_isna_old`, please let me know how I can run the performance tests