@joy-rosie 2019-08-03
Hello, I ran the performance tests as shown in the link and this is the final output:

.       before           after         ratio
     [143bc34a]       [7aef72ce]
     <master>         <issue-27482-isnull>
+      21.7±0.5ms       29.6±0.6ms     1.37  indexing.NumericSeriesIndexing.time_loc_scalar(<class 'pandas.core.indexes.numeric.UInt64Index'>, 'nonunique_monotonic_inc')
+         409±5μs         498±10μs     1.22  indexing.NumericSeriesIndexing.time_ix_scalar(<class 'pandas.core.indexes.numeric.Int64Index'>, 'nonunique_monotonic_inc')
+     12.4±0.09ms       15.0±0.1ms     1.21  index_object.SetOperations.time_operation('int', 'symmetric_difference')
+      6.12±0.2ms       6.99±0.3ms     1.14  stat_ops.SeriesMultiIndexOps.time_op(0, 'median')
+     1.67±0.06μs       1.87±0.1μs     1.12  index_cached_properties.IndexCache.time_values('PeriodIndex')
+       332±0.6ns         369±10ns     1.11  timestamp.TimestampOps.time_tz_localize('US/Eastern')
+     1.01±0.02μs      1.12±0.02μs     1.11  index_cached_properties.IndexCache.time_inferred_type('RangeIndex')
+     2.14±0.05ms      2.37±0.04ms     1.11  reindex.LevelAlign.time_align_level
+         329±2ns          362±9ns     1.10  timestamp.TimestampOps.time_tz_localize(tzutc())
+       331±0.5ns          365±8ns     1.10  timestamp.TimestampOps.time_tz_localize(<UTC>)
-      4.64±0.2μs      4.21±0.05μs     0.91  timedelta.TimedeltaConstructor.time_from_missing
-     1.71±0.02μs      1.55±0.05μs     0.91  period.PeriodProperties.time_property('M', 'year')
-     1.42±0.08ms      1.29±0.01ms     0.90  ctors.SeriesConstructors.time_series_constructor(<function list_of_lists>, False, 'int')
-        567±50μs        511±0.9μs     0.90  groupby.GroupByMethods.time_dtype_as_field('float', 'head', 'transformation')
-      1.04±0.4ms          936±3μs     0.90  groupby.GroupByMethods.time_dtype_as_group('int', 'cumsum', 'direct')
-     1.68±0.01μs       1.51±0.1μs     0.90  period.PeriodProperties.time_property('min', 'second')
-     1.48±0.08ms      1.34±0.02ms     0.90  ctors.SeriesConstructors.time_series_constructor(<function list_of_lists>, True, 'int')
-     2.41±0.05ms      2.16±0.09ms     0.90  reindex.LevelAlign.time_reindex_level
-     1.69±0.04μs      1.52±0.06μs     0.90  period.PeriodProperties.time_property('M', 'month')
-     1.69±0.02μs      1.51±0.06μs     0.89  period.PeriodProperties.time_property('M', 'day')
-        586±60μs        523±0.7μs     0.89  groupby.GroupByMethods.time_dtype_as_field('datetime', 'tail', 'transformation')
-     1.69±0.01μs      1.49±0.06μs     0.88  period.PeriodProperties.time_property('min', 'day')
-      1.43±0.2ms         1.24±0ms     0.87  groupby.GroupByMethods.time_dtype_as_group('int', 'rank', 'transformation')
-        609±90μs        528±0.9μs     0.87  groupby.GroupByMethods.time_dtype_as_group('int', 'quantile', 'transformation')
-      1.46±0.2ms         1.24±0ms     0.85  groupby.GroupByMethods.time_dtype_as_group('int', 'rank', 'direct')
-      7.26±0.1ms       6.01±0.3ms     0.83  algorithms.Duplicated.time_duplicated('first', 'int')
-        348±80μs          287±1μs     0.83  groupby.GroupByMethods.time_dtype_as_field('float', 'last', 'transformation')
-      1.69±0.3ms      1.36±0.01ms     0.81  groupby.GroupByMethods.time_dtype_as_field('int', 'value_counts', 'direct')
-       90.4±20μs       65.4±0.1μs     0.72  algorithms.SortIntegerArray.time_argsort(1000)
-       422±200μs        289±0.9μs     0.68  groupby.GroupByMethods.time_dtype_as_field('float', 'last', 'direct')
-       887±300μs          553±1μs     0.62  groupby.GroupByMethods.time_dtype_as_field('object', 'head', 'transformation')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.