@joy-rosie 2019-08-03
I feel like it is a bit random when the performance has gone down and where its gone up