@jbrockmendel 2019-08-03
> I feel like it is a bit random where the performance has gone down and where its gone up

Yah, this is a common issue with asv (especially in cases where there are no non-noise changes).  The usual solution is to re-run it a couple of times and pick out the changes that are consistent across runs.