@jreback 2019-08-04
this also needs a whatsnew note, put in 1.0 in Missing bug fix section