@jbrockmendel 2019-08-07
> sorry to be a pain but where do i add a whatsnew note?

Don't worry about it, we were all new here at one point.  whatsnew note goes in doc/source/whatsnew/0.25.1.rst

> I am also finding it difficult to get any reliable results for the performance tests, I have ran it a few times and there are different results each time

That usually indicates that there isn't any real impact.  The non-asv alternative is to use Ipython %timeit directly on the affected function(s)