@joy-rosie 2019-08-13
> That usually indicates that there isn't any real impact. The non-asv alternative is to use Ipython %timeit directly on the affected function(s)
for benchmarking I am running the test function:

```
import pandas as pd
from pandas.util import testing as tm
isna_f = pd.isna
pd.__version__
```

```
%%timeit -n 10000

assert not isna_f(1.0)
assert isna_f(None)
assert isna_f(pd.np.nan)
assert not isna_f(pd.np.inf)
assert not isna_f(-pd.np.inf)

# # type
# assert not isna_f(type(pd.Series()))
# assert not isna_f(type(pd.DataFrame()))

# series
for s in [
    tm.makeFloatSeries(),
    tm.makeStringSeries(),
    tm.makeObjectSeries(),
    tm.makeTimeSeries(),
    tm.makePeriodSeries(),
]:
    assert isinstance(isna_f(s), pd.Series)

# frame
for df in [
    tm.makeTimeDataFrame(),
    tm.makePeriodFrame(),
    tm.makeMixedDataFrame(),
]:
    result = isna_f(df)
    expected = df.apply(isna_f)
    tm.assert_frame_equal(result, expected)
```

Outputs:

old:
```
'0.25.0
```
```
33 ms ± 1.74 ms per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

new:
```
'0.25.0+71.g7aef72ced'
```
```
33 ms ± 1.8 ms per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Please let me know if you think I should run the benchmark on anything else

> 
> Don't worry about it, we were all new here at one point. whatsnew note goes in doc/source/whatsnew/0.25.1.rst

Will this do into bugfixes or missing?