@jbrockmendel 2019-08-13
> Will this go into bugfixes or missing?

Yes. Probably in the "Missing" section