@joy-rosie 2019-08-14
I added into whatsnew and it seems like there are some things failing but I am not sure why a change in whatsnew could cause these particular failures?